pyproject.toml
[project]
name = "donor-db-builder"
version = "0.1.0"
description = ""
authors = [
    {name = "Torin Rozzelle"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "neo4j",
    "pandas",
    "loguru",
    "python-dotenv",
    "numpy",
    "requests (>=2.32.3,<3.0.0)",
    "polars (>=1.19.0,<2.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "duckdb",
    "duckdb-engine (>=0.14.2,<0.15.0)",
    "httpx",
    "trafilatura (>=2.0.0,<3.0.0)",
    "pyarrow (>=18.1.0,<19.0.0)",
    "spacy (>=3.8.3,<4.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "llama-index (>=0.12.10,<0.13.0)",
    "llama-index-embeddings-huggingface (>=0.5.0,<0.6.0)",
    "llama-index-vector-stores-duckdb (>=0.3.0,<0.4.0)",
    "llama-index-llms-ollama (>=0.5.0,<0.6.0)",
    "ipywidgets (>=8.1.5,<9.0.0)",
    "llama-index-readers-wikipedia (>=0.3.0,<0.4.0)",
    "dlt (>=1.5.0,<2.0.0)",
    "typer (>=0.15.1,<0.16.0)",
    "rich (>=13.9.4,<14.0.0)",
    "textual (>=1.0.0,<2.0.0)",
    "click-shell (>=2.1,<3.0)",
    "pyte (>=0.8.2,<0.9.0)",
    "cyclopts (>=3.2.0,<4.0.0)",
    "textual-dev (>=1.7.0,<2.0.0)",
    "textual-fastdatatable (>=0.11.0,<0.12.0)",
    "orjson (>=3.10.14,<4.0.0)",
    "textual-autocomplete==2.1.0b0",
]

packages = [
    "src/donor_db_builder"
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.1",
]

[project.scripts]
raising = "donor_db_builder.__main__:run"

[tool.hatch.version]
path = "src/donor_db_builder/__init__.py"
